storage test added auth context is added login and register page is refactored to have form seperately
