+++ src/auth/auth-context.types.ts
@@ -1,0 +1,5 @@
+import { makeTypes } from 'common/common-helper';
+
+export const auth = {
+  ...makeTypes('SIGN_IN'),
+};
